requirements.txt
matplotlib==3.8.4
mrcfile==1.5.0
pandas==2.2.1
pytorch-lightning==1.8.0.post1
PyYAML==6.0.1
scikit-image==0.22.0
scipy==1.13.0
tqdm==4.65.0
typer==0.12.0
typer-cli==0.12.0
typer-config==1.4.0
typer-slim==0.12.0
typing_extensions==4.9.0